split shelf into pres and cont, implemented props for books/shelfs
