--- phoenix/components/container/Book.tsx
@@ -1,11 +1,0 @@
-import React from 'react'
-import { View } from 'react-native'
-
-interface BookProps {
-    title: string;
-    cover?: string;
-    author?: string;
-    publicationDate: Date;
-}
-
-export const BookContainer = (props: BookProps) => {}
