+++ phoenix/components/presenter/ShelfPresenter.tsx
@@ -1,0 +1,13 @@
+import React from "react"
+import { View, FlatList, Text } from "react-native"
+
+const ShelfPresenter = () => {
+    return (
+        <View>
+            <FlatList
+                data={[]}
+                renderItem={() => <Text></Text>}
+            />
+        </View>
+    )
+}
